reorganized scripts, updated autosetup, and changed the setup script to delegate nft for the 0th address
