+++ scripts/goveranance-deployment-script.js
@@ -2,8 +2,7 @@ const hre = require("hardhat");
 async function main() {
     const MyGovernor = await ethers.getContractFactory("Governance");
     console.log("Deploying MyGovernor");
-    const myGovernor = await MyGovernor.deploy("0x5FbDB2315678afecb367f032d93F642f64180aa3");
-    // possible cause of bug in next line?
+    const myGovernor = await MyGovernor.deploy("0x5FbDB2315678afecb367f032d93F642f64180aa3", 'UDAOGovnor');
     const deployedGovernor  = await myGovernor.deployed()
     console.log("MyGovernor deployed to:", myGovernor.address);
   }
@@ -14,3 +13,4 @@ main()
     console.error(error);
     process.exit(1);
   });
+
