+++ scripts/membershipNFT-deployment-script.js
@@ -1,21 +1,11 @@
-
-const hre = require("hardhat");
-
 async function main() {
 
   const MembershipNFT = await hre.ethers.getContractFactory("MembershipNFT");
-  const membershipNFT = await MembershipNFT.deploy();
+  const membershipNFT = await MembershipNFT.deploy('UDAOToken', 'UDAO');
 
   await membershipNFT.deployed();
-
   console.log("MembershipNFT deployed to:", membershipNFT.address);
   
-  const UDAOToken = await hre.ethers.getContractFactory("UDAOToken");
-  const udaotoken = await UDAOToken.deploy();
-
-  await udaotoken.deployed();
-
-  console.log("UDAOToken deployed to:", udaotoken.address);
 }
